+++ spec/unit/honeybadger/util/revision_spec.rb
@@ -18,4 +18,14 @@ describe Honeybadger::Util::Revision do
     ENV['HEROKU_SLUG_COMMIT'] = 'heroku revision'
     expect(Honeybadger::Util::Revision.detect).to eq('heroku revision')
   end
+
+  it "returns nil when detected value is a blank string" do
+    allow(Honeybadger::Util::Revision).to receive(:from_git).and_return(" ")
+    expect(Honeybadger::Util::Revision.detect).to eq(nil)
+  end
+
+  it "returns nil when detected value is nil" do
+    allow(Honeybadger::Util::Revision).to receive(:from_git).and_return(nil)
+    expect(Honeybadger::Util::Revision.detect).to eq(nil)
+  end
 end